pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "personal-stem-instructor"
version = "0.1.0"
description = "Personalized STEM tutoring agent with RAG, course planning, and assessments."
authors = [
  { name = "Henry", email = "henry@example.com" }
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pydantic>=2.6.0",
  "pyyaml>=6.0.1",
  "tqdm>=4.66.0",
  "numpy>=1.24.0",
  "scikit-learn>=1.3.0",
  "sentence-transformers>=2.6.0",
  "chromadb>=0.4.24",
  "faiss-cpu>=1.7.4",
  "pymupdf>=1.22.0",
  "markdown-it-py>=3.0.0",
  "python-dotenv>=1.0.0",
  "typer>=0.9.0",
  "rich>=13.7.0",
  "structlog>=24.1.0",
  "openai-agents>=0.2.1"
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4.0",
  "pytest-asyncio>=0.21.0",
  "black>=24.1.0",
  "ruff>=0.3.0",
  "mypy>=1.8.0"
]

[project.scripts]
ai-tutor = "ai_tutor.cli:app"

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.ruff]
line-length = 88
select = ["E", "F", "B", "I"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q"
testpaths = [
  "tests",
  "src"
]
